+++ src/sisl/io/__init__.py
@@ -141,7 +141,7 @@ VASP
    eigenvalSileVASP
    chgSileVASP
    locpotSileVASP
-   outSileVASP
+   stdoutSileVASP
 
 
 Wannier90