+++ CHANGELOG.md
@@ -8,6 +8,7 @@ we hit release version 1.0.0.
 ## [X.Y.Z] - YYYY-MM-DD
 
 ### Added
+- added `read_trajectory` to read cell vectors, atomic positions, and forces from VASP OUTCAR
 - slicing io files multiple output (still WIP), see #584 for details
   Intention is to have all methods use this method for returning
   multiple values, it should streamline the API.
@@ -43,6 +44,8 @@ we hit release version 1.0.0.
 - `BrillouinZone.merge` allows simple merging of several objects, #537
 
 ### Changed
+- `stdoutSileVASP` will not accept `all=` arguments
+- `stdoutSileVASP.read_energy` returns as default the next item (no longer the last)
 - `txtSileOrca` will not accept `all=` arguments, see #584
 - `stdoutSileOrca` will not accept `all=` arguments, see #584
 - `xyzSile` out from sisl will now default to the extended xyz file-format